# apigrpc/apigrpc.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Nakama API v2",
    "version": "2.0",
    "contact": {
      "name": "The Nakama Authors \u0026 Contributors",
      "url": "https://github.com/heroiclabs/nakama",
      "email": "hello@heroiclabs.com"
    }
  },
  "host": "127.0.0.1:7350",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v2/rpc/{id}": {
      "get": {
        "summary": "Execute a Lua function on the server.",
        "operationId": "RpcFunc2",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiRpc"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "The identifier of the function.",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "payload",
            "description": "The payload of the function which must be a JSON object.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "http_key",
            "description": "The authentication key used when executed as a non-client HTTP request.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Nakama"
        ],
        "security": [
          {
            "BearerJwt": [],
            "HttpKeyAuth": []
          }
        ]
      },
      "post": {
        "summary": "Execute a Lua function on the server.",
        "operationId": "RpcFunc",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiRpc"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "The identifier of the function.",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "description": "The payload of the function which must be a JSON object.",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Nakama"
        ],
        "security": [
          {
            "BearerJwt": [],
            "HttpKeyAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "apiRpc": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "The identifier of the function."
        },
        "payload": {
          "type": "string",
          "description": "The payload of the function which must be a JSON object."
        },
        "http_key": {
          "type": "string",
          "description": "The authentication key used when executed as a non-client HTTP request."
        }
      },
      "description": "Execute an Lua function on the server."
    },
...
  },
  "securityDefinitions": {
    "BasicAuth": {
      "type": "basic"
    },
    "BearerJwt": {
      "type": ""
    },
    "HttpKeyAuth": {
      "type": "apiKey",
      "name": "http_key",
      "in": "header"
    }
  },
  "security": [
    {
      "BearerJwt": []
    }
  ],
...
}